JWT Claims and Authorization

+++ resources/item.py
@@ -2,7 +2,7 @@ import uuid
 from flask import request
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt
 
 from db import db
 from models import ItemModel
@@ -16,18 +16,22 @@ blp = Blueprint("Items", "items", description="Operations on items")
 
 @blp.route("/item/<int:item_id>")
 class Item(MethodView):
-    @jwt_required
+    @jwt_required()
     @blp.response(200, ItemSchemas)
     def get(self, item_id):
         item = ItemModel.query.get_or_404(item_id)
         return item
 
-    @jwt_required
+    @jwt_required()
     def delete(self, item_id):
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege required.")
+
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
-        return {"message": "Item deleted"}
+        return {"message": "Item deleted."}
 
     @blp.arguments(ItemUpdateSchema)
     @blp.response(200, ItemSchemas)
